requirements.txt
numpy==1.25.1
pygame==2.5.1
PyGLM==2.7.0
PyOpenGL==3.1.7
